+++ Source/Model/Conversation/ZMConversation+Ephemeral.swift
@@ -166,22 +166,6 @@ extension TimeInterval {
     }
 }
 
-//TODO: new file
-public extension ZMConversation {
-    @NSManaged internal var language: String
-
-    public var inputLanguage: String? {
-        get {
-            return language
-        }
-
-        set {
-            if let newValue = newValue {
-                language = newValue
-            }
-        }
-    }
-}
 
 public extension ZMConversation {
