+++ Tests/Source/Model/Observer/ConversationObserverTests.swift
@@ -52,7 +52,8 @@ class ConversationObserverTests : NotificationDispatcherTestBase {
             "securityLevelChanged",
             "createdRemotelyChanged",
             "allowGuestsChanged",
-            "destructionTimeoutChanged"
+            "destructionTimeoutChanged",
+            "languageChanged"
         ]
     }
     
@@ -475,18 +476,35 @@ class ConversationObserverTests : NotificationDispatcherTestBase {
     func testThatLocalDestructionTimeoutChangeIsTriggeringObservation()
     {
         // given
-        let conversation = ZMConversation.insertNewObject(in:self.uiMOC)
-        conversation.conversationType = ZMConversationType.group
-        uiMOC.saveOrRollback()
-        XCTAssert(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
-        
+        let conversation = setupGroupConversation()
+
         // when
         self.checkThatItNotifiesTheObserverOfAChange(conversation,
                                                      modifier: { conversation, _ in conversation.localMessageDestructionTimeout = 1000 },
                                                      expectedChangedField: "destructionTimeoutChanged",
                                                      expectedChangedKeys: [#keyPath(ZMConversation.localMessageDestructionTimeout)])
     }
-    
+
+    func setupGroupConversation() -> ZMConversation{
+        let conversation = ZMConversation.insertNewObject(in:self.uiMOC)
+        conversation.conversationType = ZMConversationType.group
+        uiMOC.saveOrRollback()
+        XCTAssert(waitForAllGroupsToBeEmpty(withTimeout: 0.5))
+
+        return conversation
+    }
+
+    func testThatLanguageChangeIsTriggeringObservation() {
+        // given
+        let conversation = setupGroupConversation()
+
+        // when
+        self.checkThatItNotifiesTheObserverOfAChange(conversation,
+                                                     modifier: { conversation, _ in conversation.language = "de-DE" },
+                                                     expectedChangedField: "languageChanged",
+                                                     expectedChangedKeys: [#keyPath(ZMConversation.language)])
+    }
+
     func testThatAccessRoleChangeIsTriggeringObservation()
     {
         // given
